refactor: cmp naming typo
