+++ frontend/src/lib/components/proposal-detail/ProposalSystemInfoSection.svelte
